requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.2
deprecation==2.1.0
Flask==2.2.2
functions-framework==3.2.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-core==2.3.2
google-cloud-secret-manager==2.12.6
google-cloud-storage==2.5.0
google-cloud-vision==3.1.3
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gphotospy==0.2.1
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.50.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
piexif==1.1.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
zipp==3.9.0
